requirements.txt
PySide6
pydantic
sqlalchemy
python-dotenv